@@ -1,15 +1,14 @@
 {
   "name": "getkirby/cms",
-  "type": "kirby-cms",
   "description": "The Kirby 3 core",
+  "license": "proprietary",
+  "type": "kirby-cms",
+  "version": "3.6.2",
   "keywords": [
     "kirby",
     "cms",
     "core"
   ],
-  "homepage": "https://getkirby.com",
-  "version": "3.6.1.1",
-  "license": "proprietary",
   "authors": [
     {
       "name": "Kirby Team",
@@ -17,32 +16,30 @@
       "homepage": "https://getkirby.com"
     }
   ],
+  "homepage": "https://getkirby.com",
+  "support": {
+    "email": "support@getkirby.com",
+    "issues": "https://github.com/getkirby/kirby/issues",
+    "forum": "https://forum.getkirby.com",
+    "source": "https://github.com/getkirby/kirby"
+  },
+  "_comment": "TODO: psr/log is not used by Kirby; drop pinned version when Kirby no longer supports PHP 7",
   "require": {
     "php": ">=7.4.0 <8.2.0",
     "ext-ctype": "*",
     "ext-mbstring": "*",
     "claviska/simpleimage": "3.6.5",
-    "filp/whoops": "2.14.4",
+    "filp/whoops": "2.14.5",
     "getkirby/composer-installer": "^1.2.1",
     "laminas/laminas-escaper": "2.9.0",
     "michelf/php-smartypants": "1.8.1",
-    "mustangostang/spyc": "0.6.3",
     "phpmailer/phpmailer": "6.5.3",
     "psr/log": "1.1.4",
-    "symfony/polyfill-intl-idn": "1.23.0",
-    "symfony/polyfill-mbstring": "1.23.1"
+    "symfony/polyfill-intl-idn": "1.24.0",
+    "symfony/polyfill-mbstring": "1.24.0"
   },
   "replace": {
     "symfony/polyfill-php72": "*"
-  },
-  "config": {
-    "optimize-autoloader": true,
-    "platform-check": false
-  },
-  "extra": {
-    "unused": [
-      "symfony/polyfill-intl-idn"
-    ]
   },
   "autoload": {
     "psr-4": {
@@ -54,6 +51,18 @@
     "files": [
       "config/setup.php",
       "config/helpers.php"
+    ]
+  },
+  "config": {
+    "allow-plugins": {
+      "getkirby/composer-installer": true
+    },
+    "optimize-autoloader": true,
+    "platform-check": false
+  },
+  "extra": {
+    "unused": [
+      "symfony/polyfill-intl-idn"
     ]
   },
   "scripts": {
@@ -77,12 +86,5 @@
     "fix": "php-cs-fixer fix",
     "test": "phpunit --stderr --coverage-html=tests/coverage",
     "zip": "composer archive --format=zip --file=dist"
-  },
-  "support": {
-    "email": "support@getkirby.com",
-    "issues": "https://github.com/getkirby/kirby/issues",
-    "forum": "https://forum.getkirby.com",
-    "source": "https://github.com/getkirby/kirby"
-  },
-  "_comment": "TODO: psr/log is not used by Kirby; drop pinned version when Kirby no longer supports PHP 7"
+  }
 }
