@@ -7,27 +7,13 @@ use Kirby\Cms\Language;
  */
 return [
     'fields' => [
-        'code' => function (Language $language) {
-            return $language->code();
-        },
-        'default' => function (Language $language) {
-            return $language->isDefault();
-        },
-        'direction' => function (Language $language) {
-            return $language->direction();
-        },
-        'locale' => function (Language $language) {
-            return $language->locale();
-        },
-        'name' => function (Language $language) {
-            return $language->name();
-        },
-        'rules' => function (Language $language) {
-            return $language->rules();
-        },
-        'url' => function (Language $language) {
-            return $language->url();
-        },
+        'code'      => fn (Language $language) => $language->code(),
+        'default'   => fn (Language $language) => $language->isDefault(),
+        'direction' => fn (Language $language) => $language->direction(),
+        'locale'    => fn (Language $language) => $language->locale(),
+        'name'      => fn (Language $language) => $language->name(),
+        'rules'     => fn (Language $language) => $language->rules(),
+        'url'       => fn (Language $language) => $language->url(),
     ],
     'type'  => 'Kirby\Cms\Language',
     'views' => [
