@@ -7,18 +7,10 @@ use Kirby\Cms\Role;
  */
 return [
     'fields' => [
-        'description' => function (Role $role) {
-            return $role->description();
-        },
-        'name' => function (Role $role) {
-            return $role->name();
-        },
-        'permissions' => function (Role $role) {
-            return $role->permissions()->toArray();
-        },
-        'title' => function (Role $role) {
-            return $role->title();
-        },
+        'description' => fn (Role $role) => $role->description(),
+        'name'        => fn (Role $role) => $role->name(),
+        'permissions' => fn (Role $role) => $role->permissions()->toArray(),
+        'title'       => fn (Role $role) => $role->title(),
     ],
     'type'  => 'Kirby\Cms\Role',
     'views' => [
