@@ -7,37 +7,17 @@ use Kirby\Form\Form;
  * Site
  */
 return [
-    'default' => function () {
-        return $this->site();
-    },
+    'default' => fn () => $this->site(),
     'fields' => [
-        'blueprint' => function (Site $site) {
-            return $site->blueprint();
-        },
-        'children' => function (Site $site) {
-            return $site->children();
-        },
-        'content' => function (Site $site) {
-            return Form::for($site)->values();
-        },
-        'drafts' => function (Site $site) {
-            return $site->drafts();
-        },
-        'files' => function (Site $site) {
-            return $site->files()->sorted();
-        },
-        'options' => function (Site $site) {
-            return $site->permissions()->toArray();
-        },
-        'previewUrl' => function (Site $site) {
-            return $site->previewUrl();
-        },
-        'title' => function (Site $site) {
-            return $site->title()->value();
-        },
-        'url' => function (Site $site) {
-            return $site->url();
-        },
+        'blueprint'  => fn (Site $site) => $site->blueprint(),
+        'children'   => fn (Site $site) => $site->children(),
+        'content'    => fn (Site $site) => Form::for($site)->values(),
+        'drafts'     => fn (Site $site) => $site->drafts(),
+        'files'      => fn (Site $site) => $site->files()->sorted(),
+        'options'    => fn (Site $site) => $site->permissions()->toArray(),
+        'previewUrl' => fn (Site $site) => $site->previewUrl(),
+        'title'      => fn (Site $site) => $site->title()->value(),
+        'url'        => fn (Site $site) => $site->url(),
     ],
     'type' => 'Kirby\Cms\Site',
     'views' => [
