@@ -7,20 +7,11 @@ use Kirby\Cms\SiteBlueprint;
  */
 return [
     'fields' => [
-        'name' => function (SiteBlueprint $blueprint) {
-            return $blueprint->name();
-        },
-        'options' => function (SiteBlueprint $blueprint) {
-            return $blueprint->options();
-        },
-        'tabs' => function (SiteBlueprint $blueprint) {
-            return $blueprint->tabs();
-        },
-        'title' => function (SiteBlueprint $blueprint) {
-            return $blueprint->title();
-        },
+        'name'    => fn (SiteBlueprint $blueprint) => $blueprint->name(),
+        'options' => fn (SiteBlueprint $blueprint) => $blueprint->options(),
+        'tabs'    => fn (SiteBlueprint $blueprint) => $blueprint->tabs(),
+        'title'   => fn (SiteBlueprint $blueprint) => $blueprint->title(),
     ],
-    'type' => 'Kirby\Cms\SiteBlueprint',
-    'views' => [
-    ],
+    'type'  => 'Kirby\Cms\SiteBlueprint',
+    'views' => [],
 ];
