@@ -8,37 +8,17 @@ use Kirby\Toolkit\Str;
  */
 return [
     'fields' => [
-        'ascii' => function () {
-            return Str::$ascii;
-        },
-        'authStatus' => function () {
-            return $this->kirby()->auth()->status()->toArray();
-        },
-        'defaultLanguage' => function () {
-            return $this->kirby()->panelLanguage();
-        },
-        'isOk' => function (System $system) {
-            return $system->isOk();
-        },
-        'isInstallable' => function (System $system) {
-            return $system->isInstallable();
-        },
-        'isInstalled' => function (System $system) {
-            return $system->isInstalled();
-        },
-        'isLocal' => function (System $system) {
-            return $system->isLocal();
-        },
-        'multilang' => function () {
-            return $this->kirby()->option('languages', false) !== false;
-        },
-        'languages' => function () {
-            return $this->kirby()->languages();
-        },
-        'license' => function (System $system) {
-            return $system->license();
-        },
-        'locales' => function () {
+        'ascii'           => fn () => Str::$ascii,
+        'authStatus'      => fn () => $this->kirby()->auth()->status()->toArray(),
+        'defaultLanguage' => fn () => $this->kirby()->panelLanguage(),
+        'isOk'            => fn (System $system) => $system->isOk(),
+        'isInstallable'   => fn (System $system) => $system->isInstallable(),
+        'isInstalled'     => fn (System $system) => $system->isInstalled(),
+        'isLocal'         => fn (System $system) => $system->isLocal(),
+        'multilang'       => fn () => $this->kirby()->option('languages', false) !== false,
+        'languages'       => fn () => $this->kirby()->languages(),
+        'license'         => fn (System $system) => $system->license(),
+        'locales'         => function () {
             $locales = [];
             $translations = $this->kirby()->translations();
             foreach ($translations as $translation) {
@@ -46,21 +26,11 @@ return [
             }
             return $locales;
         },
-        'loginMethods' => function (System $system) {
-            return array_keys($system->loginMethods());
-        },
-        'requirements' => function (System $system) {
-            return $system->toArray();
-        },
-        'site' => function (System $system) {
-            return $system->title();
-        },
-        'slugs' => function () {
-            return Str::$language;
-        },
-        'title' => function () {
-            return $this->site()->title()->value();
-        },
+        'loginMethods' => fn (System $system) => array_keys($system->loginMethods()),
+        'requirements' => fn (System $system) => $system->toArray(),
+        'site'         => fn (System $system) => $system->title(),
+        'slugs'        => fn () => Str::$language,
+        'title'        => fn () => $this->site()->title()->value(),
         'translation' => function () {
             if ($user = $this->user()) {
                 $translationCode = $user->language();
@@ -74,12 +44,8 @@ return [
                 return $this->kirby()->translation('en');
             }
         },
-        'kirbytext' => function () {
-            return $this->kirby()->option('panel.kirbytext') ?? true;
-        },
-        'user' => function () {
-            return $this->user();
-        },
+        'kirbytext' => fn () => $this->kirby()->option('panel.kirbytext') ?? true,
+        'user' => fn () => $this->user(),
         'version' => function () {
             $user = $this->user();
 
