@@ -7,21 +7,11 @@ use Kirby\Cms\Translation;
  */
 return [
     'fields' => [
-        'author' => function (Translation $translation) {
-            return $translation->author();
-        },
-        'data' => function (Translation $translation) {
-            return $translation->dataWithFallback();
-        },
-        'direction' => function (Translation $translation) {
-            return $translation->direction();
-        },
-        'id' => function (Translation $translation) {
-            return $translation->id();
-        },
-        'name' => function (Translation $translation) {
-            return $translation->name();
-        },
+        'author'    => fn (Translation $translation) => $translation->author(),
+        'data'      => fn (Translation $translation) => $translation->dataWithFallback(),
+        'direction' => fn (Translation $translation) => $translation->direction(),
+        'id'        => fn (Translation $translation) => $translation->id(),
+        'name'      => fn (Translation $translation) => $translation->name(),
     ],
     'type'  => 'Kirby\Cms\Translation',
     'views' => [
