@@ -7,58 +7,24 @@ use Kirby\Form\Form;
  * User
  */
 return [
-    'default' => function () {
-        return $this->user();
-    },
+    'default' => fn () => $this->user(),
     'fields' => [
-        'avatar' => function (User $user) {
-            return $user->avatar() ? $user->avatar()->crop(512) : null;
-        },
-        'blueprint' => function (User $user) {
-            return $user->blueprint();
-        },
-        'content' => function (User $user) {
-            return Form::for($user)->values();
-        },
-        'email' => function (User $user) {
-            return $user->email();
-        },
-        'files' => function (User $user) {
-            return $user->files()->sorted();
-        },
-        'id' => function (User $user) {
-            return $user->id();
-        },
-        'language' => function (User $user) {
-            return $user->language();
-        },
-        'name' => function (User $user) {
-            return $user->name()->value();
-        },
-        'next' => function (User $user) {
-            return $user->next();
-        },
-        'options' => function (User $user) {
-            return $user->panel()->options();
-        },
-        'panelImage' => function (User $user) {
-            return $user->panel()->image();
-        },
-        'permissions' => function (User $user) {
-            return $user->role()->permissions()->toArray();
-        },
-        'prev' => function (User $user) {
-            return $user->prev();
-        },
-        'role' => function (User $user) {
-            return $user->role();
-        },
-        'roles' => function (User $user) {
-            return $user->roles();
-        },
-        'username' => function (User $user) {
-            return $user->username();
-        }
+        'avatar'      => fn (User $user) => $user->avatar() ? $user->avatar()->crop(512) : null,
+        'blueprint'   => fn (User $user) => $user->blueprint(),
+        'content'     => fn (User $user) => Form::for($user)->values(),
+        'email'       => fn (User $user) => $user->email(),
+        'files'       => fn (User $user) => $user->files()->sorted(),
+        'id'          => fn (User $user) => $user->id(),
+        'language'    => fn (User $user) => $user->language(),
+        'name'        => fn (User $user) => $user->name()->value(),
+        'next'        => fn (User $user) => $user->next(),
+        'options'     => fn (User $user) => $user->panel()->options(),
+        'panelImage'  => fn (User $user) => $user->panel()->image(),
+        'permissions' => fn (User $user) => $user->role()->permissions()->toArray(),
+        'prev'        => fn (User $user) => $user->prev(),
+        'role'        => fn (User $user) => $user->role(),
+        'roles'       => fn (User $user) => $user->roles(),
+        'username'    => fn (User $user) => $user->username()
     ],
     'type'  => 'Kirby\Cms\User',
     'views' => [
