@@ -3,6 +3,7 @@
 namespace Kirby\Cms;
 
 use Kirby\Http\Response;
+use Kirby\Toolkit\I18n;
 use Whoops\Handler\CallbackHandler;
 use Whoops\Handler\Handler;
 use Whoops\Handler\PlainTextHandler;
@@ -136,7 +137,7 @@ trait AppErrors
                     'status'  => 'error',
                     'code'    => $code,
                     'details' => $details,
-                    'message' => 'An unexpected error occurred! Enable debug mode for more info: https://getkirby.com/docs/reference/system/options/debug',
+                    'message' => I18n::translate('error.unexpected'),
                 ], $httpCode);
             }
 
