@@ -49,6 +49,6 @@ class KirbyTags
 
                 return $match[0];
             }
-        }, $text);
+        }, $text ?? '');
     }
 }
