@@ -121,7 +121,7 @@ class SmartyPants
     public function parse(string $text = null): string
     {
         // prepare the text
-        $text = str_replace('&quot;', '"', $text);
+        $text = str_replace('&quot;', '"', $text ?? '');
 
         // parse the text
         return $this->parser->transform($text);
